LPD-89734 Rename editor config contributor samples to -1 and -2
