+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-1/client-extension.dev.yaml
@@ -1,2 +1,2 @@
 liferay-sample-editor-config-contributor-1:
-    baseURL: http://localhost:3001/build
+    baseURL: http://localhost:3000/build
