+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-1/client-extension.yaml
@@ -3,7 +3,12 @@ assemble:
         into: static
 liferay-sample-editor-config-contributor-1:
     editorConfigKeys:
-        -   sampleReactCKEditor5ClassicEditor
-    name: Liferay Sample Editor Config Contributor CKEditor 5 WordCount
+        -   description
+        -   fragmentEntryLinkRichTextEditor
+        -   sampleAlloyEditor
+        -   sampleClassicEditor
+        -   sampleLegacyEditor
+        -   sampleReactClassicEditor
+    name: Liferay Sample Editor Config Contributor
     type: editorConfigContributor
     url: index.js
