+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-1/src/index.ts
@@ -3,39 +3,68 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {WordCount} from '@ckeditor/ckeditor5-word-count/dist/index.js';
 import {
 	EditorConfigTransformer,
 	EditorTransformer,
 } from '@liferay/js-api/editor';
 
 const editorConfigTransformer: EditorConfigTransformer<any> = (config) => {
-	let displayEl: HTMLElement | null = null;
+
+	// Alloy Editor
+
+	const toolbars: any = config.toolbars;
+
+	if (typeof toolbars === 'object') {
+		interface ISelection {
+			buttons: Array<string>;
+			name: string;
+		}
+
+		const textSelection: ISelection = toolbars.styles?.selections?.find(
+			(selection: ISelection) => selection.name === 'text'
+		);
+
+		if (textSelection.buttons) {
+			textSelection.buttons.push('video');
+
+			return {
+				...config,
+				toolbars,
+			};
+		}
+	}
+
+	// CKEditor
+
+	const toolbar: string | [string[]] = config.toolbar;
+
+	const buttonName = 'AICreator';
+	let transformedConfig: any;
+
+	if (typeof toolbar === 'string') {
+		const activeToolbar = config[`toolbar_${toolbar}`];
+
+		activeToolbar.push([buttonName]);
+
+		transformedConfig = {
+			...config,
+			[`toolbar_${toolbar}`]: activeToolbar,
+		};
+	}
+	else if (Array.isArray(toolbar)) {
+		toolbar.push([buttonName]);
+
+		transformedConfig = {
+			...config,
+			toolbar,
+		};
+	}
+
+	const extraPlugins: string = config.extraPlugins;
 
 	return {
-		...config,
-		extraPlugins: [...(config.extraPlugins ?? []), WordCount],
-		wordCount: {
-			onUpdate: ({
-				characters,
-				words,
-			}: {
-				characters: number;
-				words: number;
-			}) => {
-				if (!displayEl) {
-					displayEl = document.createElement('div');
-					displayEl.className = 'mt-1 text-secondary';
-
-					document
-						.querySelector('.ck-editor__editable')
-						?.closest('.ck-editor')
-						?.parentElement?.appendChild(displayEl);
-				}
-
-				displayEl.textContent = `Words: ${words} | Characters: ${characters}`;
-			},
-		},
+		...transformedConfig,
+		extraPlugins: extraPlugins ? `${extraPlugins},aicreator` : 'aicreator',
 	};
 };
