+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-1/src/types.d.ts
