+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-2/package.json
@@ -1,0 +1,18 @@
+{
+	"description": "Liferay Sample Editor Config Contributor CKEditor 5 WordCount",
+	"devDependencies": {
+		"@ckeditor/ckeditor5-word-count": "46.0.3",
+		"@liferay/js-api": "0.6.2",
+		"serve": "^14.2.0",
+		"typescript": "^5.0.4"
+	},
+	"main": "./src/index.ts",
+	"name": "@liferay/liferay-sample-editor-config-contributor-2",
+	"scripts": {
+		"build": "esbuild src/index.ts --outdir=build --bundle --format=esm --external:@ckeditor/ckeditor5-word-count",
+		"start": "PORT=3001 serve --cors",
+		"watch": "tsc --watch"
+	},
+	"type": "module",
+	"version": "0.0.0"
+}
