+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor-2/tsconfig.json
