--- workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor/client-extension.dev.yaml
@@ -1,2 +1,0 @@
-liferay-sample-editor-config-contributor:
-    baseURL: http://localhost:3000/build
