--- workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor/client-extension.yaml
@@ -1,14 +1,0 @@
-assemble:
-    -   from: build
-        into: static
-liferay-sample-editor-config-contributor:
-    editorConfigKeys:
-        -   description
-        -   fragmentEntryLinkRichTextEditor
-        -   sampleAlloyEditor
-        -   sampleClassicEditor
-        -   sampleLegacyEditor
-        -   sampleReactClassicEditor
-    name: Liferay Sample Editor Config Contributor
-    type: editorConfigContributor
-    url: index.js
