--- workspaces/liferay-sample-workspace/client-extensions/liferay-sample-editor-config-contributor/package.json
@@ -1,17 +1,0 @@
-{
-	"description": "Liferay Sample Editor Config Contributor",
-	"devDependencies": {
-		"@liferay/js-api": "0.6.2",
-		"serve": "^14.2.0",
-		"typescript": "^5.0.4"
-	},
-	"main": "./src/index.ts",
-	"name": "@liferay/liferay-sample-editor-config-contributor",
-	"scripts": {
-		"build": "esbuild src/index.ts --outdir=build --bundle --format=esm",
-		"start": "PORT=3000 serve --cors",
-		"watch": "tsc --watch"
-	},
-	"type": "module",
-	"version": "0.0.0"
-}
