+++ workspaces/liferay-sample-workspace/package.json
@@ -17,8 +17,8 @@
 			"client-extensions/liferay-sample-custom-element-2",
 			"client-extensions/liferay-sample-custom-element-4",
 			"client-extensions/liferay-sample-custom-element-5",
-			"client-extensions/liferay-sample-editor-config-contributor",
 			"client-extensions/liferay-sample-editor-config-contributor-1",
+			"client-extensions/liferay-sample-editor-config-contributor-2",
 			"client-extensions/liferay-sample-etc-node",
 			"client-extensions/liferay-sample-fds-cell-renderer",
 			"client-extensions/liferay-sample-fds-filter",
